[Fix]: changed scores to be stored as floats

+++ pages/api/adjudications/edit.js
@@ -55,9 +55,9 @@ export default async (req, res) => {
       id: id,
     },
     data: {
-      artistic_mark: parseInt(artisticMark),
-      technical_mark: parseInt(technicalMark),
-      cumulative_mark: parseInt(cumulativeMark),
+      artistic_mark: parseFloat(artisticMark),
+      technical_mark: parseFloat(technicalMark),
+      cumulative_mark: parseFloat(cumulativeMark),
       audio_url: audioUrl,
       notes: notes,
     },
